pyproject.toml
[tool.poetry]
name = "testhillelgit"
version = "0.1.0"
description = ""
authors = ["mildrein <“jagma2008@gmail.com”>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
